% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_connect.R
\name{mdl_connect}
\alias{mdl_connect}
\title{Get connection to Moodle database}
\usage{
mdl_connect(type, host, port, name, user, passw)
}
\arguments{
\item{type}{connection type for database}

\item{host}{ip or host address}

\item{port}{database port number}

\item{name}{database name}

\item{user}{user name}

\item{passw}{password}
}
\value{
A dbplyr reference object.
}
\description{
Returns a pool object; basically a wrapper for dbPool
}
