% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_course_objects.R
\name{mdl_course_objects}
\alias{mdl_course_objects}
\title{Returns infos on all objects in a Moodle course}
\usage{
mdl_course_objects(dbp, prefix, courseid)
}
\arguments{
\item{dbp}{a pool object for db connection}

\item{prefix}{prefix for tables, e.g. "m_"}

\item{courseid}{the unique course ID of the course whose log data is being retrieved}
}
\value{
A dataframe with the name of all objects in a Moodle course
}
\description{
Returns a dataframe names of all objects in a Moodle course
}
