% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_course_objects.R
\name{mdl_course_objects_query}
\alias{mdl_course_objects_query}
\title{Raw query to retrieve names of all objects in a Moodle course}
\usage{
mdl_course_objects_query(dbp, prefix, courseid)
}
\arguments{
\item{dbp}{a pool object for db connection}

\item{prefix}{prefix for tables, e.g. "m_"}

\item{courseid}{the unique course ID of the course whose log data is being retrieved}
}
\value{
Character vector with SQL query
}
\description{
Returns a SQL query to retrieve names of all objects in a Moodle course.
}
