% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_courses.R
\name{mdl_courses}
\alias{mdl_courses}
\title{Get Courses}
\usage{
mdl_courses(dbp, prefix)
}
\arguments{
\item{dbp}{a pool object for db connection}

\item{prefix}{prefix for tables, e.g. "m_"}
}
\value{
A dataframe with the course list.
}
\description{
Returns a reference to the (cached) course table, with the most relevant columns selected.
}
\details{
For convenience a join with the category table is made, and "category_name" added
}
