% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_forum.R
\name{mdl_forumdata}
\alias{mdl_forumdata}
\title{Provides Moodle forum data from the specified course}
\usage{
mdl_forumdata(dbp, prefix, courseid, forumids)
}
\arguments{
\item{dbp}{a pool object for db connection}

\item{prefix}{prefix for tables, e.g. "m_"}

\item{courseid}{the unique course ID of the course whose log data is being retrieved}

\item{forumids}{vector of forum ids, if present, function only returns messages from these forums}
}
\value{
A dataframe with the forum data, restricted to...
}
\description{
Returns a dataframe with Moodle forum data for the specified course.
}
