% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_forum.R
\name{mdl_forumdata_query}
\alias{mdl_forumdata_query}
\title{Raw query for Moodle forum data}
\usage{
mdl_forumdata_query(prefix, courseid, forumids)
}
\arguments{
\item{prefix}{prefix for tables, e.g. "m_"}

\item{courseid}{the unique course ID of the course whose log data is being retrieved}

\item{forumids}{vector of forum ids, if present, function only returns messages from these forums}
}
\value{
Character vector with SQL query
}
\description{
Returns a SQL query to retrieve Moodle log data for the specified course.
}
