% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_forum_list.R
\name{mdl_forumlist}
\alias{mdl_forumlist}
\title{Provides a data frame with all forums in the specified course}
\usage{
mdl_forumlist(dbp, prefix, courseid)
}
\arguments{
\item{dbp}{a pool object for db connection}

\item{prefix}{prefix for tables, e.g. "m_"}

\item{courseid}{the unique course ID of the course whose log data is being retrieved}
}
\value{
A dataframe with the forum data, restricted to...
}
\description{
Returns a dataframe with Moodle forum data for the specified course.
}
