% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_forum_list.R
\name{mdl_forumlist_query}
\alias{mdl_forumlist_query}
\title{Raw query for Moodle forum data}
\usage{
mdl_forumlist_query(prefix, courseid)
}
\arguments{
\item{prefix}{prefix for tables, e.g. "m_"}

\item{courseid}{the unique course ID of the course whose log data is being retrieved}
}
\value{
Character vector with SQL query
}
\description{
Returns a SQL query to retrieve a list with all forums from a Moodle course
}
