% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_groups.R
\name{mdl_grouplist}
\alias{mdl_grouplist}
\title{Provides a list of all groups of a Moodle course}
\usage{
mdl_grouplist(dbp, prefix, courseid)
}
\arguments{
\item{dbp}{a pool object for db connection}

\item{prefix}{prefix for tables, e.g. "m_"}

\item{courseid}{the unique course ID of the course whose data is being retrieved}
}
\value{
A dataframe a list of all groups
}
\description{
Returns a dataframe with a list of groups for the specified course.
}
