% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_groups.R
\name{mdl_groupslist_query}
\alias{mdl_groupslist_query}
\title{Raw query to retrieve all groups within a Moodle course}
\usage{
mdl_groupslist_query(prefix, courseid)
}
\arguments{
\item{prefix}{prefix for tables, e.g. "m_"}

\item{courseid}{the unique course ID of the course whose data is being retrieved}
}
\value{
Character vector with SQL query
}
\description{
Returns a SQL query to retrieve all groups and their groupings for the specified course.
}
