% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_log.R
\name{mdl_logdata}
\alias{mdl_logdata}
\title{Provides Moodle log data for the specified period}
\usage{
mdl_logdata(dbp, prefix, courseid, startstamp, stopstamp)
}
\arguments{
\item{dbp}{a pool object for db connection}

\item{prefix}{prefix for tables, e.g. "m_"}

\item{courseid}{the unique course ID of the course whose log data is being retrieved}

\item{startstamp}{Unix timestamp that marks the beginning of the time period from which the log data is being retrieved}

\item{stopstamp}{Unix timestamp that marks the end of the time period from which the log data is being retrieved}
}
\value{
A dataframe with the logdata, restricted to tbd
}
\description{
Returns a dataframe with Moodle log data for the specified course and time period.
}
