% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_log.R
\name{mdl_logdata_query}
\alias{mdl_logdata_query}
\title{Raw query for Moodle log data}
\usage{
mdl_logdata_query(prefix, courseid, startstamp, stopstamp)
}
\arguments{
\item{prefix}{prefix for tables, e.g. "m_"}

\item{courseid}{the unique course ID of the course whose log data is being retrieved}

\item{startstamp}{Unix timestamp that marks the beginning of the time period from which the log data is being retrieved}

\item{stopstamp}{Unix timestamp that marks the end of the time period from which the log data is being retrieved}
}
\value{
Character vector with SQL query
}
\description{
Returns a SQL query to retrieve Moodle log data for the specified course and time period.
}
