% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_resolve_objecttable.R
\name{mdl_resolve_objectids}
\alias{mdl_resolve_objectids}
\title{Resolves all objecttable and objectids in the provided Moodle logdata}
\usage{
mdl_resolve_objectids(dbp, prefix, logdata)
}
\arguments{
\item{dbp}{a pool object for db connection}

\item{prefix}{prefix for tables, e.g. "m_"}

\item{logdata}{dataframe with Moodle logdata}
}
\value{
A dataframe with Moodle logdata with additional column objectname
}
\description{
Returns the dataframe with Moodle logdata with additional column objectname
}
