% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_user.R
\name{mdl_userlist}
\alias{mdl_userlist}
\title{Provides a list of all users of a Moodle course}
\usage{
mdl_userlist(dbp, prefix, courseid, returngroups = FALSE)
}
\arguments{
\item{dbp}{a pool object for db connection}

\item{prefix}{prefix for tables, e.g. "m_"}

\item{courseid}{the unique course ID of the course whose data is being retrieved}

\item{returngroups}{if true, return all groups and groupings for each user}
}
\value{
A dataframe a list of all users
}
\description{
Returns a dataframe with a list of all users (student role) for the specified course.
}
