% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{moodleLA}
\alias{moodleLA}
\title{moodleLA app}
\usage{
moodleLA()
}
\description{
Returns a Shiny app to start using shiny::runApp. See examples.
}
\examples{
# The following line starts the app inside the systems standard browser on port 3000.
\dontrun{
  shiny::runApp(moodleLA(), port = 3000, launch.browser = TRUE)
}
}
