% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presidio.R
\name{presidio_language_supported}
\alias{presidio_language_supported}
\title{Checks whether or not Microsoft Presidio supports the requested language}
\usage{
presidio_language_supported(
  language = "en",
  analyzer_url = "http://localhost:5002/analyze"
)
}
\arguments{
\item{language}{Language code (e.g. \code{"en"}, \code{"de"}).}

\item{analyzer_url}{Base URL of the Analyzer service \emph{without} trailing slash.
Default \code{"http://localhost:5002"}.}
}
\value{
Logical indicating support of requested language
}
\description{
Sends the input text(s) to Presidio \emph{Analyzer} and pipes the result
directly into Presidio \emph{Anonymizer}.
Works with a single string or a character vector.
}
