% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudonymise.R
\name{pseudonymize}
\alias{pseudonymize}
\alias{pseudonymise}
\title{Pseudonymise data}
\usage{
pseudonymize(input, salt)

pseudonymise(input, salt)
}
\arguments{
\item{input}{a single value or a vector}

\item{salt}{a random value added to the input before hashing}
}
\value{
Pseudonymised result based on input and salt, either a single value or a vector.
}
\description{
Pseudonymises data such as usernames or groupnames.
}
