% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudonymise.R
\name{pseudonymize_matricle}
\alias{pseudonymize_matricle}
\alias{pseudonymise_matricle}
\title{Pseudonymise matricle number}
\usage{
pseudonymize_matricle(matricle, salt)

pseudonymise_matricle(matricle, salt)
}
\arguments{
\item{matricle}{a single value}

\item{salt}{a random value added to the input before hashing}
}
\value{
A pseudonymised value based on input and salt.
}
\description{
Pseudonymises data such as student matricle number.
}
