% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_forum.R
\name{pseudonymize_text}
\alias{pseudonymize_text}
\alias{pseudonymise_text}
\title{Remove personal information from a text}
\usage{
pseudonymize_text(
  text,
  remove_email = TRUE,
  remove_url = TRUE,
  remove_phonenumber = TRUE,
  placeholder = TRUE
)

pseudonymise_text(
  text,
  remove_email = TRUE,
  remove_url = TRUE,
  remove_phonenumber = TRUE,
  placeholder = TRUE
)
}
\arguments{
\item{text}{A character from which to remove personal information}

\item{remove_email}{Remove email addresses from text}

\item{remove_url}{Remove URLs from text}

\item{remove_phonenumber}{Remove phone numbers from text}

\item{placeholder}{Shall the personal information be replaced by placeholders (EMAIL, URL, PHONE)}
}
\value{
Returns the input with personal information removed.
}
\description{
Helper function to remove email addresses, URLs, and phone numbers from text.
}
