% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_forum.R
\name{remove_names}
\alias{remove_names}
\title{Remove names from text}
\usage{
remove_names(text, firstnames, lastnames)
}
\arguments{
\item{text}{Either a single character/string or character vector from which to remove personal information}

\item{firstnames}{lastnames A character vector of names, same length as lastnames}

\item{lastnames}{A character vector of names, same length as firstnames}
}
\value{
Text with names replaces with \if{html}{\out{<PERSON>}}
}
\description{
Function names from a text and replace it with \if{html}{\out{<PERSON>}}
}
